Add note about failures, and rename for convention
