+++ lib/routes/api/metadata/processor.js
@@ -7,7 +7,7 @@
 const debug = require('debug')('magnet-metadata-service:processor');
 const htmlParser = require('magnet-html-parser');
 const request = require('./request');
-const { getDataForUrl } = require('../../../processors/contentservice');
+const { getDataForUrl } = require('../../../processors/content-service');
 
 const types = {
   json: 'application/json',
@@ -65,7 +65,9 @@ module.exports = function(url, options) {
           return Object.assign(metadata, res.extra_metadata || {});
         })
         .catch((err) => {
-          // Ignore
+          // Ignore, if the fetch to the content service fails for any reason,
+          // don't fail the entire response. The additional data from the content
+          // isn't mandatory.
           return metadata;
         })
         .then((metadata) => {
